+++ jkorm-spring/src/test/kotlin/com/tang/jkorm/spring/entity/Account.kt
@@ -1,6 +1,6 @@
 package com.tang.jkorm.spring.entity
 
-import com.tang.jkorm.annotation.Id
+import com.tang.jkorm.annotation.id.Id
 
 /**
  * @author Tang