+++ jkorm/src/main/kotlin/com/tang/jkorm/annotation/id/strategy/snowflake/SnowflakeIdStrategy.kt
@@ -1,13 +1,13 @@
-package com.tang.jkorm.utils.id.defaults
+package com.tang.jkorm.annotation.id.strategy.snowflake
 
-import com.tang.jkorm.utils.id.IdStrategy
+import com.tang.jkorm.annotation.id.strategy.IdStrategy
 import com.tang.jkorm.utils.id.Ids
 import java.lang.reflect.Field
 
 /**
  * @author Tang
  */
-class DefaultIdStrategy : IdStrategy {
+class SnowflakeIdStrategy : IdStrategy {
 
     override fun getId(idField: Field): Any {
         return Ids.snowflake().let { if (idField.type == String::class.java) it.toString() else it }